pep8 docstrings, move ExpBackoff to separate file
