+++ environment.yaml
@@ -3,7 +3,5 @@ channels:
   - defaults
 dependencies:
   - python=3.7
-  - pip
-  - pip:
-      - requests
-      - distributed
+  - requests
+  - distributed
